// source/ucp-portal/ucp-react/src/models/iconName.ts
// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
// SPDX-License-Identifier: Apache-2.0

export enum IconName {
    REMOVE = 'remove',
    EDIT = 'edit',
    ADD = 'add-plus',
    STATUS_NEGATIVE = 'status-negative',
    USER_PROFILE_ACTIVE = 'user-profile-active',
    STATUS_INFO = 'status-info',
    REFRESH = 'refresh',
    FLAG = 'flag',
    GEN_AI = 'gen-ai',
}